fix: vayras help in unit test
